gv_cinema_app/gv_cinema_app/common/pagination.py
from rest_framework.response import Response
from rest_framework.pagination import PageNumberPagination


class ApiResultsSetPagination(PageNumberPagination):
    page_size = 25
    page_size_query_param = 'page_size'
    max_page_size = 200

    def get_paginated_response(self, data):
        return Response({
            'count': self.page.paginator.count,
            'num_pages': self.page.paginator.num_pages,
            'results': data
        })
